Documentation and refactor
The derive macro has been refactored to use results rather than quote_spanned which is a bit more neat
