+++ lerp_derive/Cargo.toml
@@ -16,6 +16,7 @@ edition = "2018"
 [dependencies]
 syn = { version = "1.0", features = ["full"] }
 quote = "1.0"
+proc-macro2 = "1.0"
 
 [lib]
 proc-macro = true
